Migrate float arguments to double arguments.
- Adds CRTOProductViewEvent tests
- Adds serializer tests for product view event
- Adds serializer tests for product list view event
- Refactors JSON serializer tests for size and readability

JIRA: MOB-477

+++ advertiser-sdk/advertiser-sdk/product/CRTOBasketProduct.h
@@ -15,8 +15,8 @@
 /** A string containing the product ID used to initialize the receiver. (read-only) */
 @property (nonatomic,strong,readonly) NSString* productId;
 
-/** A float value representing the approximate price used to initialize the receiver. (read-only) */
-@property (nonatomic,readonly) float price;
+/** A double value representing the approximate price used to initialize the receiver. (read-only) */
+@property (nonatomic,readonly) double price;
 
 /** A signed integer value representing the quantity of product used to initialize the receiver. (read-only) */
 @property (nonatomic,readonly) NSInteger quantity;
@@ -32,13 +32,13 @@
  *  Initializes a newly allocated instance of @c CRTOBasketProduct with a product ID, a price, and a quanity.
  *
  *  @param productId A string containing the product ID of the product. This value must match the product ID supplied to Criteo in your organization's catalog feed.
- *  @param price     The unit price of the product, represented as a single-precision floating-point value.
+ *  @param price     The unit price of the product, represented as a double-precision floating-point value.
  *  @param quantity  An integer representing the number of units of product.
  *
  *  @note The loss of precision associated with representing price as a floating-point value does not impact Criteo's ability to process your event data.
  *
  *  @return A @c CRTOBasketProduct initialized with a product ID, a price, and a quantity.
  */
-- (instancetype) initWithProductId:(NSString*)productId price:(float)price quantity:(NSInteger)quantity;
+- (instancetype) initWithProductId:(NSString*)productId price:(double)price quantity:(NSInteger)quantity;
 
 @end
